Feat: 오브젝트 생성 로직, 자동 클릭 구현
1. 오브젝트가 생성이 되고 삭제 시 계속 생성 됨
2. 마우스로 오브젝트 클릭 시 스프라이트의 a값이 증가하고 1이 되면 삭제되게 함
3. 클릭할 때마다 해당 오브젝트의 골드를 획득 함
4. 자동으로 클릭 되게 구현

+++ Assets/ScriptableObject/ObjectData.cs
@@ -7,6 +7,6 @@ public class ObjectData : ScriptableObject
 {
     public string objectName; // 물건 이름
     public Sprite sprite;     // 이미지
-    public int count;         // 횟수
+    public float count;         // 횟수
     public int gold;          // 골드 량
 }
